feat(ai.notebooks): fix pr comments
