+++ packages/manager/apps/pci-ai-notebooks/src/pages/Root.layout.tsx
@@ -6,22 +6,17 @@ import {
   useParams,
 } from 'react-router-dom';
 import { useRouting, useShell } from '@ovh-ux/manager-react-shell-client';
-
 import { useEffect } from 'react';
 import { defineCurrentPage } from '@ovh-ux/request-tagger';
 import queryClient from '@/query.client';
-
 import { useLoadingIndicatorContext } from '@/contexts/LoadingIndicator.context';
 import { getProject } from '@/data/api/project/project.api';
 import Breadcrumb from '@/components/breadcrumb/Breadcrumb.component';
 import BreadcrumbItem from '@/components/breadcrumb/BreadcrumbItem.component';
-
 import { Toaster } from '@/components/ui/toaster';
 import PageLayout from '@/components/page-layout/PageLayout.component';
-import Auth from './auth/auth.page';
 import { UserActivityProvider } from '@/contexts/UserActivityContext';
 import { USER_INACTIVITY_TIMEOUT } from '@/configuration/polling.constants';
-import { useGetAuthorization } from '@/hooks/api/ai/authorization/useGetAuthorization.hook';
 import { useTrackPageAuto } from '@/hooks/useTracking';
 
 export function breadcrumb() {
@@ -85,29 +80,12 @@ export function useNotebooksData() {
 }
 
 export default function Layout() {
-  const { projectId } = useParams();
-  const authorizationQuery = useGetAuthorization(projectId);
-  if (authorizationQuery.isSuccess && authorizationQuery.data.authorized) {
-    return (
-      <PageLayout>
-        <UserActivityProvider timeout={USER_INACTIVITY_TIMEOUT}>
-          <Breadcrumb />
-          <RoutingSynchronisation />
-          <Outlet />
-          <Toaster />
-        </UserActivityProvider>
-      </PageLayout>
-    );
-  }
   return (
     <PageLayout>
       <UserActivityProvider timeout={USER_INACTIVITY_TIMEOUT}>
+        <Breadcrumb />
         <RoutingSynchronisation />
-        <Auth
-          onSuccess={() => {
-            authorizationQuery.refetch();
-          }}
-        />
+        <Outlet />
         <Toaster />
       </UserActivityProvider>
     </PageLayout>
