+++ packages/manager/apps/pci-ai-notebooks/src/pages/notebooks/[notebookId]/attached-data/_components/VolumesListTable.component.tsx
@@ -22,11 +22,7 @@ export default function VolumesList({ volumes }: Readonly<VolumesListProps>) {
     },
   });
 
-  return (
-    <>
-      <DataTable columns={columns} data={volumes} pageSize={25} />
-    </>
-  );
+  return <DataTable columns={columns} data={volumes} pageSize={25} />;
 }
 
 VolumesList.Skeleton = function VolumesListSkeleton() {
