+++ packages/manager/apps/pci-ai-notebooks/src/pages/notebooks/[notebookId]/backups/_components/BackupsListTable.component.tsx
@@ -18,11 +18,7 @@ export default function BackupsList({ backups }: Readonly<BackupsListProps>) {
     },
   });
 
-  return (
-    <>
-      <DataTable columns={columns} data={backups} pageSize={25} />
-    </>
-  );
+  return <DataTable columns={columns} data={backups} pageSize={25} />;
 }
 
 BackupsList.Skeleton = function BackupsListSkeleton() {
